requirements.txt
Django==1.6.5
South==1.0
cffi==0.8
cogapp==2.3
colorama==0.3.1
django-social-auth==0.7.28
fakeredis==0.4.3
greenlet==0.4.0
httplib2==0.9
mock==1.0.1
nose==1.3.3
oauth2==1.5.211
pinocchio==0.4.1
python-openid==2.2.5
pytz==2014.4
readline==6.2.4.1
redis==2.10.3
wsgiref==0.1.2
